Fix sidebar z-index
Due to the sidebar's z-index only being set to 10 when the sidebar-toggle is active, and unset when inactive, the sidebar appears to be below the other components of the main page content when sliding out of view. I fixed this by applying the z-index property to .sidebar regardless of whether sidebar-toggle is active or not.